Create MPU6050_SAMD51_XFFT.ino
Let's say I want to use the MPU6050 with an SAMD51 microcontroller. How would I go about reading the MPU6050 data, writing that data to a buffer or accumulator and then performing a second set of operations on that data such as a Fast Fourier transform? 

The process of reading the MPU6050 data and storing it in a buffer to perform further operations such as Fast Fourier Transform (FFT) involves a few steps.

Firstly, you need to make sure the MPU6050 library you are using supports the SAMD51 microcontroller and the communication interface (likely I2C) is properly configured.

Secondly, you will need to set up the MPU6050 and begin reading data from it. This data will be acceleration and gyro data in x, y, and z axes.

Then, you need to store this data in a buffer. Depending on the type of data you want to process with FFT, you may need to transform or preprocess the raw data from the MPU6050.

Once you have enough data in your buffer, you can perform an FFT. The ArduinoFFT library can be used for this purpose.

Here is a simplified code snippet demonstrating how to read MPU6050 data, store it into a buffer, and perform an FFT:

Please note that this code reads only the x-axis acceleration data from MPU6050. If you need to use other axes or gyro data, you should modify it accordingly. This code also assumes the MPU6050 library provides the function readRawAccel() to get acceleration data.

Also, you'll need to adjust the SAMPLES and SAMPLING_FREQUENCY parameters to match your specific use case. The number of samples should be a power of 2, and the sampling frequency should be at least twice the maximum frequency you expect to measure (according to the Nyquist-Shannon theorem).

Remember to install the ArduinoFFT library (you can find it in the Arduino Library Manager) before compiling the code.