Create MPU6050_DLPF_Butter.ino
To adjust the Digital Low-Pass Filter (DLPF) using a Butterworth filter, you would need to implement the Butterworth filter algorithm, which can be quite complex and requires a good understanding of digital signal processing.

However, it's important to note that the MPU6050 already uses a built-in hardware DLPF which is likely a Butterworth filter or a similar design, and it's unclear how you would combine this with an additional software-based Butterworth filter.

That said, if you wanted to apply a second-order Butterworth filter to the accelerometer readings in software, you could use a library like the "Filter" library available in the Arduino Library Manager. Here's a very simplified example:

In this example, we apply a Butterworth filter to the accelerometer magnitude before using it to adjust the DLPF. This provides an additional level of smoothing to the signal, which might be useful if the raw accelerometer data is very noisy.

Please note that you'll need to install the Filter library to your Arduino IDE. You can do this by going to Sketch -> Include Library -> Manage Libraries and searching for "Filter". Select the one by "Alexander M. West".

Remember to install and include any necessary libraries when using third-party or custom filtering algorithms, and always test your code thoroughly to ensure that the filter is performing as expected. Also note that performing digital filtering in software will consume more of your microcontroller's processing time and may increase the delay between sensor readings.