Create MemoryUsageInfoAVR.ino
This program uses the special symbols __heap_start, __brkval, __data_start, and _end that are defined by the linker to give the locations of different regions of memory.

The getFreeRam function computes the amount of free RAM by subtracting the top of the heap (either __heap_start or __brkval, depending on whether __brkval is non-zero) from the current stack pointer (represented by the local variable v).

The stack size is computed by subtracting the top of the heap from the current stack pointer.

The sum of the sizes of the .data and .bss sections is computed by subtracting the start of the .data section from the end of the .bss section (represented by _end).

Please note that the actual amount of free RAM may be less than the value returned by getFreeRam if the stack and heap have overlapped due to excessive memory usage.

You could name the file MemoryUsageInfo.ino. Remember that Arduino IDE requires the filename to match the main sketch name.

Please ensure you have the right board and port selected in your Arduino IDE before uploading this program.

Also note, these methods might not work for all architectures, and might need some tweaking or different approaches depending on the specific board you're using. It's always important to refer to the documentation of your board to see how to get this kind of information.

Please note, this code assumes you're using an AVR-based or similar microcontroller. If you're using an ARM-based board (like the Arduino Due or SAMD51), this code might not work, and you'll need to use the tools and techniques specific to those boards to get this kind of information. As of my last training data in September 2021, Arduino didn't provide a built-in way to access this information for ARM-based boards. You might need to use board-specific functions or linker symbols, or possibly a third-party library. Always refer to the latest documentation and resources for your specific hardware.

This code may need to be adjusted depending on the microcontroller and compiler being used. Please verify if these symbols are available and correctly provide the desired memory usage information for your specific board and compiler.