Create MemoryUsageInfoSAMD51.ino
Here is an example of an Arduino program which can print some memory usage details for a SAMD51 microcontroller.

Please note that getting this kind of detailed memory information on ARM microcontrollers can be more complex than on AVR, and might not always be possible in the same way. You might need to use board-specific functions or linker symbols, or possibly a third-party library.

This example prints out the total amount of free heap memory. It doesn't provide the same breakdown between stack and heap as the AVR version, because the stack and heap grow towards each other, and there isn't a clear separation between them in the same way.

This code uses the FreeRTOS xPortGetFreeHeapSize and xPortGetMinimumEverFreeHeapSize functions to print the current and minimum amount of free heap memory. These are standard FreeRTOS functions, but they might not be available if you're not using a FreeRTOS-based board package, or if it's configured not to include them.

Please note that you might need to include the FreeRTOS headers to access these functions:

#include "FreeRTOS.h"
#include "task.h"

This code should work on a SAMD51 board using the Adafruit SAMD Boards package, or a similar FreeRTOS-based package. Always refer to the latest documentation and resources for your specific hardware.