Create 328P_entropy_FFT.ino
Certainly, if you're not seeing output in the serial monitor and you suspect that the Arduino Uno's microcontroller may be constrained by the code execution speed, you can slow down the process by reducing the sampling rate and the FFT size. Here's a modified version of the code with reduced parameters to better suit the limitations of the ATmega328P microcontroller on the Arduino Uno:

In this modified code, the sampling frequency has been reduced to 500 Hz, and the number of samples has been reduced to 64 to ease the computational load on the ATmega328P microcontroller. Additionally, the delay between sampling points has been increased to 2000 microseconds to slow down the process, allowing you to see the visualization in the serial monitor.